requirements.txt
Flask==3.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0